DOP-5036 refactor delete stale
